feat: :sparkles: Adicionar funcionalidade de atualização de cliente
- Adiciona método na classe ClienteService para atualizar informações de um cliente existente. - Adiciona endpoint PUT no ClienteController para atualizar um cliente por ID.

+++ src/main/java/br/com/apifatec/apifatec/domain/cliente/ClienteService.java
@@ -25,6 +25,29 @@ public class ClienteService {
 	}
 
 	public void deletarCliente(Long id) {
+
+		Cliente cliente = encontrarClientePorId(id);
+
+		if (cliente == null)
+		{
+			throw new IllegalArgumentException("Cliente não existe");
+		}	
 		clienteRepository.deleteById(id);
+
+	}
+
+		public Cliente atualizarCliente(Long id, Cliente cliente) {
+			Cliente clienteCadastradoCliente = encontrarClientePorId(id);
+		
+			if (clienteCadastradoCliente == null)
+			{
+				return null;
+			} else {
+				clienteCadastradoCliente.setNome(cliente.getNome());
+				clienteCadastradoCliente.setRazaoSocial(cliente.getRazaoSocial());
+				clienteCadastradoCliente.setEmail(cliente.getEmail());
+				clienteCadastradoCliente.setEndereco(cliente.getEndereco());
+				return clienteRepository.save(clienteCadastradoCliente);
+		}
 	}
 }
